requirements.txt
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
autopep8==1.4.4
certifi==2019.6.16
cffi==1.12.3
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-eventstream==2.5.0
django-filter==2.1.0
django-grip==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
gripcontrol==3.3.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pubcontrol==2.4.2
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
zope.interface==4.6.0
